pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "multimodal-hybrid-parsing"
version = "0.1.0"
description = "A library for multimodal hybrid document parsing"
authors = [
    {name = "George Strong", email = "geowstrong@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "docling==2.23.0",
    "pytest==7.0.0",
    "reportlab==4.2.5",
    "Pillow==10.4.0",
    "openai==1.59.7",
    "aiohttp==3.11.11",
    "tenacity==9.0.0",
    "torch==2.5.1",
    "torchvision==0.20.1"
]

[tool.setuptools]
packages = ["multimodal_hybrid_parsing"]